refactor: extract custom hook handling files attachments and smartlink
used in add-attachments-dropdown
